requirements.txt
Package                            Version
---------------------------------- -----------
aiohappyeyeballs                   2.6.1
aiohttp                            3.13.2
aiosignal                          1.4.0
alembic                            1.15.2
annotated-types                    0.7.0
anyio                              4.12.0
asgiref                            3.10.0
attrs                              25.4.0
awsebcli                           3.26
bcrypt                             4.0.1
blessed                            1.27.0
blinker                            1.9.0
botocore                           1.42.33
Bottleneck                         1.4.2
cachetools                         6.2.4
cement                             2.10.14
certifi                            2025.11.12
cffi                               2.0.0
charset-normalizer                 3.4.4
click                              8.2.1
cloudpickle                        3.1.2
colorama                           0.4.6
contourpy                          1.3.3
cryptography                       46.0.3
cycler                             0.12.1
databricks-sdk                     0.76.0
decorator                          5.2.1
Deprecated                         1.3.1
distro                             1.9.0
Django                             5.2.8
django-picklefield                 3.4.0
djangorestframework                3.16.1
docker                             7.1.0
fabric                             3.2.2
fastapi                            0.115.8
fastapi-utils                      0.8.0
filelock                           3.20.1
Flask                              3.1.2
flask-cors                         6.0.2
fonttools                          4.61.1
frozenlist                         1.8.0
gitdb                              4.0.12
GitPython                          3.1.45
google-auth                        2.45.0
graphene                           3.4.3
graphql-core                       3.2.7
graphql-relay                      3.2.0
greenlet                           3.3.0
gunicorn                           23.0.0
h11                                0.16.0
httptools                          0.7.1
huey                               2.5.5
idna                               3.11
importlib_metadata                 8.7.1
invoke                             2.2.1
ipinfo                             5.3.0
itsdangerous                       2.2.0
Jinja2                             3.1.6
jmespath                           1.1.0
joblib                             1.5.3
jsonref                            1.1.0
kiwisolver                         1.4.9
Mako                               1.3.10
markdown-it-py                     4.0.0
MarkupSafe                         3.0.3
matplotlib                         3.10.7
mdurl                              0.1.2
mkl_fft                            2.1.1
mkl_random                         1.3.0
mkl-service                        2.5.2
mlflow                             3.8.0
mlflow-skinny                      3.8.0
mlflow-tracing                     3.8.0
multidict                          6.7.0
numexpr                            2.14.1
numpy                              2.3.5
opentelemetry-api                  1.38.0
opentelemetry-proto                1.39.1
opentelemetry-sdk                  1.38.0
opentelemetry-semantic-conventions 0.59b0
orjson                             3.10.18
packaging                          24.2
pandas                             2.3.3
paramiko                           4.0.0
passlib                            1.7.4
pathspec                           0.12.1
pillow                             12.0.0
pip                                25.3
propcache                          0.4.1
protobuf                           6.33.2
psutil                             5.9.8
psycopg2-binary                    2.9.11
pyarrow                            22.0.0
pyasn1                             0.6.1
pyasn1_modules                     0.4.2
pycparser                          2.23
pydantic                           2.11.9
pydantic_core                      2.33.2
Pygments                           2.19.2
PyJWT                              2.7.0
PyMySQL                            1.1.2
PyNaCl                             1.6.2
pyparsing                          3.3.1
python-dateutil                    2.9.0.post0
python-dotenv                      1.2.1
python-multipart                   0.0.21
python-rapidjson                   1.14
pytz                               2025.2
PyYAML                             6.0.3
requests                           2.32.5
requests-file                      3.0.1
rich                               14.2.0
rsa                                4.9.1
scikit-learn                       1.8.0
scipy                              1.16.3
seaborn                            0.13.2
secure                             1.0.1
semantic-version                   2.10.0
setuptools                         80.9.0
six                                1.17.0
smmap                              5.0.2
SQLAlchemy                         2.0.45
SQLAlchemy-Utils                   0.42.1
sqlmodel                           0.0.18
sqlparse                           0.5.3
starlette                          0.45.3
termcolor                          2.5.0
threadpoolctl                      3.6.0
tldextract                         5.1.3
typing_extensions                  4.15.0
typing-inspection                  0.4.2
tzdata                             2025.3
urllib3                            1.26.20
uvicorn                            0.40.0
uvloop                             0.22.1
watchfiles                         1.1.1
wcwidth                            0.2.14
websockets                         15.0.1
Werkzeug                           3.1.4
wheel                              0.45.1
wrapt                              2.0.1
yarl                               1.22.0
zenml                              0.93.0
zipp                               3.23.0
